// src/shared/constants/enums.ts
export enum JOTFORMFieldType {
  ControlHead = "control_head",
  ControlRadio = "control_radio",
  ControlDropdown = "control_dropdown",
  ControlPhone = "control_phone",
  ControlNumber = "control_number",
  ControlTextbox = "control_textbox",
  ControlPageBreak = "control_pagebreak",
  ControlFullname = "control_fullname",
  ControlEmail = "control_email",
  ControlAddress = "control_address",
  ControlDate = "control_datetime",
  ControlFileUpload = "control_fileupload",
  ControlTextarea = "control_textarea",
  ControlButton = "control_button",
  ControlHidden = "control_hidden",
}